fbtp: Fix issue where failures in locking the riser mux is not detected
Summary:
Since the bool 'ret' was used to store the return value of mux_lock() it will never be less than 0.
Thus we fail to detect if the function failed.
PS: This was a compiler warning detected by gcc 6.

Test Plan: Load on fbtp and ensure no abnormal cases are observed.

Reviewed By: benwei13

Differential Revision: D5844797

fbshipit-source-id: c55d165