Fix REST API bug (not asynchronous): for common routines only (applies to all boards)
Summary:
Fix REST API not async bug: for common routines and CMM specific routines

aoihttp assumes all the REST API handlers are implemented in "asyncio" way. That is, it needs to be defined with "async" keyword, and when a blocking function (or time consuming function) is called "await" keyword needs to be used.

Fundamental problem is that, the c functions we use to query H/W status is not designed this way.

In order to deal with this kind of situations, we use loop.run_in_executor method, which basically turns all blocking calls into asynchronous call, by spawning another thread to do the job and give the execution control back to the main job pool (loop)

The decorator in the diff takes any function as an argument, and run it in run_in_executor method in order to make it "asynchronous".

With this diff, all the REST API handler now uses this decorator, which make them asynchronous (non-blocking and concurrent)

Test Plan: Tested and verified on fsw001-p030/34/37/41/46/52.f01.ftw2

Reviewed By: cooperlees

Differential Revision: D5830468

fbshipit-source-id: e9c6fb4