fw-util: Allow only one instance of fw-util per FRU
Summary:
Prevent issues which can occur if two or more instances of fw-util are active
(especially for update) for the same FRU.

Test Plan:
Load on fw-util and try to perform fw-util update from two
parallel ssh sessions. The first should be in progress, the second should
fail with an error message.

Reviewed By: benwei13

Differential Revision: D6054822

fbshipit-source-id: a92f806