vboot: Existance of KEK enables software enforcement
Summary:
Verified boot is enforced by a R/O CS0 (hardware) or previously the existance of the `verify` variable (software).
This changes the software enforcement to also be turned on when any signing keys are present in the ROM SPL.

Since these keys are added after building an image we infer the intent of using a verified boot.
This means any error with the TPM or subordinate keys will result in booting the recovery.

Test Plan:
Build and flash to ROM SPL and BMC.
Remove the TPM and check that the recovery environment was booted.

Reviewed By: jmehta

Differential Revision: D5489196

fbshipit-source-id: 133de9d