healthd: add support for panic_on_oom
Summary: panic_on_oom causes the BMC kernel to panic which the out-of-memory issue arises. Based on the json config, you can enable or disable this feature. Also for Bryce Canyon, reboot the BMC when the Memory utilization crosses 95%

Test Plan:
Test 1:
=======
Confirmed that the flag is set correctly for fbttn:
root@rtptest2054-oob:~# cat /proc/sys/vm/panic_on_oom
1

Test 2:
=======
Set reboot threshold to 75% and write to a file to meet the threshold 75%

1. Check if system reboots when mem utilization crosses threshold: `PASS`
2. Check if the log exist: `PASS`

  healthd: Rebooting BMC; latest uptime: 723 sec

Reviewed By: amithash

Differential Revision: D6256578

fbshipit-source-id: a01f089