vboot: Set recovery flag from vbs command on error
Summary:
The `vbs` command in U-Boot can set the VBS structure's error type and error code fields.
This is used during verification of the kernel and ramdisk.

If the command is used to set non-success errors and the machine is using a software or hardware enforce then the it will go into recovery.
We should set the correct VBS structure flags for reporting.

Test Plan:
Build, and update CS1 with a failing kernel or ramdisk, but correctly sign U-Boot.
This will trigger a [6, 60] error and cause the verified U-Boot in CS1 to boot the recovery kernel.

Differential Revision: D6501078

fbshipit-source-id: 5defa54