vboot: Add hwlock check to enable HW enforce OTP at runtime
Summary:
This removes the ability to set the OTP for HW enforcement using a preprocessor define.
It moves the decision to OTP the enforcement into runtime by including a `/hwlock` property.

If a ROM (SPI0.0) in R/W mode is ever upgraded to include this property, and includes certificates,
then the ROM will disable R/W mode going forward.

We will add the following to the vboot documentation:
```
./tests/verified-boot/signing/fit-sign --lock /path/to/flash-MACHINE /tmp/output/flash-MACHINE-lockbit
```

The output flash will need to be signed afterward.

Test Plan: This needs testing on a machine with the WP breakouts.

Reviewed By: jmehta, amithash

Differential Revision: D6506219

fbshipit-source-id: 2d002b8