vboot: wdt: Move alt-boot reset into platform initialization
Summary:
This moves the WDT alt-boot restore from C into ASM.
This is needed because the SPL can grow to 84kB but we need the restore code to always be in the first 64kB.
To assure that the C runtime/linkage does not have any code or data dependencies beyond 64kB we place the check and set in the platform initialization.

This is always placed at the beginning of the SPL.

Test Plan:
Create a 3.1 image and flash to ROM.
Copy the 3.1 flash to a 3.1.1 staging location.
Build a new flash with this commit and place the first 64kB onto the 3.1.1 staging.
Flash the staged flash to BMC.

You may also create a 3.1.1 using the same method.

Reviewed By: amithash

Differential Revision: D6531095

fbshipit-source-id: f535c85