sensor-correction: Simplify mapping of table name to index
Summary:
The older method of getting the table index given the table name was ugly and confusing.
Rework it by including a name in the table which adds additional data but considerably simplifies the search.
This sets us up to be able to remove the maximum limits on tables.

Test Plan: Load on FBTP and ensure that there are not instabilities.

Reviewed By: benwei13

Differential Revision: D6826537

fbshipit-source-id: 9cee7a2