fbtp: fix receiving unexpected FRB3 event when MCERR/IERR
Summary:
BMC will check if the following three conditions fulfill or not.
1. Have KCS transaction
2. Have Postcode transcation
3. Receive Post end

If not for a period of time, BMC will sends out FRB3 event.
When MCERR/IERR is asserted, BMC will reset mb and BIOS will stop at postcode 0xE5 due to DWR.
Whne the issue happened, We received a unexpected FRB3 event because we failed at condition 1 and 2.

For condition 1, we did have kcs transactions after rebooting. However, the timestamp of /tmp/kcs_touch is same as /tmp/rst_touch.
They finish in the same second. In general, the former will be late for one second than the latter.
That is why we can not reproduce this issue all the time.

For condition 2, BIOS will boots up very fast due to DWR after mb reset. However, we store the postcode_last after BIOS booting up completed.
So we can't get the difference between current postcode and postcode_last.

To fix it, we store the postcode_last when gpioR5 is triggered. It can help us verify if BIOS boots up successfully or not after mb reset.

Reviewed By: benwei13

Differential Revision: D6892497

fbshipit-source-id: 12c5ec2