Logic to make sure rackmond is configured
Summary:
Recently we have seen cases where rackmond is not configured. This is essentially because the configuration message
we sent to rackmond was not received by rackmond. This could be timing issue (python3 is a bit slower, so it changed the
timing in everything), or IP message drop (there is no guarantee that IP message will be received.) In order to prevent
this situation from happening, I am adding a new logic to detect rackmond unconfigured state, and keep trying configuring
rackmond, until it's finally configured.

Test Plan:
1) This logic makes rackmond configured.
2) When I comment out the initial "configure_once()", rackmond is still configured eventually (which proves that the detection and retry logic works)

Reviewed By: apage43

Differential Revision: D6979552

fbshipit-source-id: 0f461d4