kernel: make defconfig change to trigger kernel rebuild
Summary:
We provide defconfig for the kernel configuration for each board. During
the kernel configuration preparation, the existing kernel.bbclass will
not copy defconfig to <build dir>/.config. That behaviors causes an
issue that when defconfig is changed, kernel is not rebuilt.

The fix uses do_configure_prepend() to copy defconfig to .config so that
any change in defconfig will trigger a kernel rebuild.

Test Plan:
Change a config in defconfig, 'bitbake linux-aspeed', check the kernel
build directory to confirm .config is updated. And check the compile log
that kernel is rebuilt.

Reviewed By: mikechoifb

Differential Revision: D7512523

fbshipit-source-id: 0403e74