Linux: fix voltage output from ast_adc
Summary:
The voltage value calculated by ast_adc is on 10mV. However, lm_sensors
is based on mV. So, need to multiple the ast_adc reading by 10.

This is applicable to both ast2400 and ast2500. Verifid on CMM that the
value reported by ast_adc for lm_sensors is 1/10th.

Test Plan: run sensors on wedge, galaxy100, and CMM.

Reviewed By: mikechoifb

Differential Revision: D7512536

fbshipit-source-id: 34dde0e