Fix fw_setenv issue on YAMP.
Summary: In Linux Kernel MTD driver, when QUAD BYTE WRITE attribute is set for a flash chip, MTD driver will assume that writes are always in 4 bytes chunk, aligned. This is not the case with our fw_setenv too, which is a u-boot tool. It writes one byte at a time, at random location. This usecase is not supported by MTD driver. So we will turn off QUAD WRITE feature, and resort to single byte write, which is the most stable / reliable

Test Plan: Works on YAMP

Reviewed By: tfangit

Differential Revision: D7747401

fbshipit-source-id: 32b9886