Linux|JTAG: Fix ast-jtag for SW mode
Summary:
There are two major issues fixed in this diff:
1. TDI and TDO share the same bit. The usual trick to read register, OR
the value w/ new value and then write back does work. For example, while
writing TMS, the reading of TDI is in fact TDO. Then when writing the
reg value back for the new TMS, the TDO value is accidentally written to
TDI. The diff stores the last value of TCK, TMS, and TDI. For each reg
write, the last value of other bits are used to construct the correct
value.
2. A bug in the base when calling simple_strtoul() in the original code.
The base was set as 1, which means only value 0 is accepted. The diff
converted it to the latest kstrtoul() call.

Test Plan:
use sysfs to write and read, comparing results with register
value and debug log.

Reviewed By: covracer

Differential Revision: D7857111

fbshipit-source-id: 233e8d4