ast-jtag: Add dummy wait after wait to ensure that the operation is complete
Summary:
Fix bug where CPLD version is displayed as:
CPLD Version: 00100000, CPLD DeviceID: FFFFFFFF
instead of:
CPLD Version: 00100000, CPLD DeviceID: 012BC043
This was a regression introduced by the change to improve ASD performance.

When ast-jtag is used in interrupt mode (Older behavior) there was sufficient
time between the interrupt firing and the data to be available. But in the new
polling mode added to enhance ASD performance this can be insufficient.
Add an extra usleep after the wait to ensure that the correct data is available.

Also refactor to move all the 'wait' logic to a single function.

Test Plan:
fw-util mb --version
<snip>
CPLD Version: 00100000, CPLD DeviceID: 012BC043
<snip>

Reviewed By: benwei13

Differential Revision: D8048807

fbshipit-source-id: 4d05781