fuzzing: Add fdt test harness to U-Boot
Summary:
This adds a fuzzing harness for libFuzzer to U-Boot.

To run this use:
```
make HOSTCC=clang HOSTCXX=clang++ test/fuzz
```

This harness should not be build with a normal `bitbake`.

Test Plan:
Use the command above to build the harness.
Run according to libFuzzer's documentation.

This is designed to be run on a devserver right now.
It is actively finding issues with the FDT/FIT parsing in U-Boot.

Reviewed By: mhlakhani

Differential Revision: D8249728

fbshipit-source-id: f704914