rest-api-v2: Defer forming path to setup
Summary:
The tree could be performed as X <- [A,B,C]. /api <- X.
With the current mechanism of forming the path when adding
the child, then we end up with /api/X /X/A, /X/B, /X/C instead
of /api/X, /api/X/A, /api/X/B, /api/X/C.

Thus defer the formation of the path to setup() when we know
for sure that the tree is fully formed.

Test Plan:
Test on FBY2 to ensure that everything works
(Where I discovered the issue in the first place).
curl http://localhost:8080/api/server1/config

Reviewed By: benwei13

Differential Revision: D8669414

fbshipit-source-id: b64672f