mterm: Request rsyslogd to reopen files after rotation
Summary:
Make mterm notify (signal) rsyslog to HUP every time it
rotates the log. This change is a NOP if there is no rsyslog
or if rsyslog is not using the mterm log in anyway. But if it
does, this requests rsyslog to reload its imfiles.

Test Plan: Overnight run with console logging in Bryce Canyon.

Differential Revision: D8941753

fbshipit-source-id: 6ae382a45