aspeed timer: do not read/write TMC38 register
Summary:
TMC38 register (Base address of Timer + 0x38) is not the interrupt mask
register in ast2400 and ast2500 chips, so it doesn't make sense to read
or write the register for interrupt control when it's aspeed chip
(count_down is true).

Test Plan: 1) booted up cmm-4.17 and everything looks normal.

Reviewed By: covracer

Differential Revision: D8961625

fbshipit-source-id: 88e5b9ef9