vboot: Fix vboot unit tests and add setup script
Summary:
This changes two expectations for the unit tests that test vboot.
Invalid keys will report general invalid firmware errors.

There is a way to emit the invalid keys error code but it is quite complicated.
A subordinate key must be replaced but maintain the existing signature.

The existing firmware signature must also not change.

This commit also adds the complete test-environment setup script.
Previously this script was kept in the conttest setup logic.

Test Plan:
1. Checkout OpenBMC
2. In the same repo, on a Ubuntu test host run `./tests/verified-boot/run.sh`
3. This takes a long time, but should return `0`

Reviewed By: covracer

Differential Revision: D9239755

fbshipit-source-id: efc79ec67