obmc-sensor: Allow poll interval larger than 256 seconds
Summary:
Change the time of the poll interval to be 32 bit to hold
larger poll intervals (earlier 255 seconds was barely good
enough). Also, use 32bits in case we want to move the poll
interval from seconds to milli-seconds (If in the future we
have certain sensors which need to be monitored at a finer
grained interval (probably for debug)).

Test Plan:
Tested by code-inspecting
(This is mostly a type change anyway.)

Reviewed By: vineelasmile

Differential Revision: D9805184

fbshipit-source-id: 523b185e6