pal: Add a common library with base stubs
Summary:
Add a common PAL stub library which can be overridden
by platform specific bbappends. This allows platform
owners to accept base defaults and not have to copy-paste
the library from other platforms.

The hope is to use this to minimize duplication in PAL
methods.
(Depends on platform diffs)

Test Plan:
1. Build fbtp image, load on platform, run a utility which depends on PAL (fan-util --get).
This ensures that the platform specific functionality is being used (and not the stubs).
All the change here is mostly build related and refactoring with no functionality touched.
So with some basic sanity tested on one platform should be sufficient.

2. Build aggregate-sensor-test and run the native executable and ensure all tests pass.

3. Build other platforms: fbtp, fby2, fbrc, yosemite, lightning, fbttn, neptune, wedge100, minipack, minilaketb

4. CIT - All builds done.

Reviewed By: vineelasmile

Differential Revision: D10178567

fbshipit-source-id: 8cc4742de