wedge_power.sh: remove unnecessary $? checks
Summary:
All uses of $? in wedge_power.sh are unnecessary: they are either used
immediately in a single if statement, and so can be more compactly
expressed by simply moving the original command inside the if; or they
are explicitly returned from functions and scripts, and so can be
removed since shell functions and scripts implicitly return the result
of the last command. Remove them.

Test Plan:
Ensure the five operations (`status`, `on`, `off`, `reset`, and
`reset -s`) still operate as expected.

Reviewed By: tfangit, mikechoifb

Differential Revision: D10320659

fbshipit-source-id: aa62696e0