wedge_power.sh: power off main power as well as userver
Summary:
wedge_power.sh is able to power on the main power during an "on"
operation, but up until now did not also power it off during an "off"
operation. Since powering off the Tomahawk for an extended period is the
only known way to resolve certain states (T29738090), change the "off"
operation to power off both main and e-COM. Also update out-of-date
usage text to reflect that the script does not only deal with userver
power anymore.

Test Plan:
Run `wedge_power.sh off` and ensure that both userver and system power
are reported as off by `wedge_power.sh status`. Then turn them back on
and ensure that the userver boots and can see the Tomahawk.

Reviewed By: mikechoifb

Differential Revision: D10320662

fbshipit-source-id: 56f1104fc