wedge_power.sh: improve logging
Summary:
Add a new shared function, try_and_log(), that reports the success or
failure of a potentially long-running operation on the command line and
to syslog. Use that function to improve and make more consistent the
logging of the "on" and "off" operations of wedge_power.sh. Also reword
some log and help messages to be more clear.

Test Plan:
Ensure `wedge_power.sh on` and `wedge_power.sh off` still work and log
their results to both stdout and syslog.

Reviewed By: mikechoifb

Differential Revision: D10349057

fbshipit-source-id: 2c6c32b21