remove "watchdog not stopped" message
Summary:
"watchdog did not stop" message is printed by kernel to alert users that
the watchdog device is not stopped at close. But I don't think such
alert is needed, and it's actually very annoying on openbmc platforms
because such warning is printed each time when watchdog is kicked (once
every a few seconds). So, let's remove the message.

Test Plan: Booted up cmm-4.17 and the warning is gone.

Reviewed By: mikechoifb

Differential Revision: D12811145

fbshipit-source-id: c8931c590