watchdog: suppress "watchdog did not stop" message
Summary:
Currently "watchdog did not stop!" alert message will be printed if
watchdog timer is not stopped at close. For example, people may see
the message when rebooting the system, or it may be logged to console
periodically if watchdog is kicked by running "echo k > /dev/watchdog"
command.

Given alert is usually used to report conditions where user's attention
is immediately required, "watchdog did not stop" is not qualified an
alert, and it's better to be removed to avoid confusions.

Test Plan: Booted up cmm-4.18 successfully.

Reviewed By: mikechoifb

fbshipit-source-id: a2265fa5a