u-boot: ensure delay between checks on RX interrupt
Summary:
If the first call to check if an interrupt was received fails, then instead of retrying
after the retry period, the code checks the register again. Fix to reorder delay
and the re-read of the ISR register.

Test Plan:
Build image based on rocko, sign and load on FBTP.
No verified boot or TPM related errors should be observed.

Reviewed By: covracer

Differential Revision: D13351950

fbshipit-source-id: 46ac85509