Add "gpio-ctrl" library
Summary:
Add "gpio-ctrl" library to "common/recipes-lib". Technically this is the
new version of "common/recipes-lib/gpio", but it's given the new name
"gpio-ctrl" mainly because "libgpio" has too many dependencies and it
will take a while to link all the dependencies with the new library.

Comparing with "libgpio_v0.1", following changes are made in "gpio-ctrl"
library:
  1) adding APIs for accessing gpio chips. This is required because gpio
     chip base number is dynamically assigned in newer kernel versions.
  2) adding "gpio_backend" layer, and all the "backend" operations are
     stored in "gpio_backend_ops" structure. Although we only support
     gpio_sysfs backend (gpio_sysfs.c) at this period, this abstract
     layer is added so we can easily/safely switched to gpio chardev
     kernel interface in the future (in case the "deprecated" gpio sysfs
     interface is removed).
  3) hiding implementation details. For example, callers doesn't know
     anything inside gpio_desc_t and gpiochip_desc_t structure. By doing
     this, library consumers will not be impacted by implementation
     changes.

Note: gpio poll related functions are not implemented yet, I (or anyone
who is interested) will take care of them later.

Test Plan:
1) ran "gpiocli" (linked with "libgpio-ctrl") on cmm (kernel 4.18):
   everything looks normal.
2) ran "gpiocli" (linked with "libgpio-ctrl") on minipack (kernel 4.1):
   everything looks normal.

Reviewed By: amithash

Differential Revision: D13567012

fbshipit-source-id: 806dd22ad