ipmid: Platform configuration the discrete sensors which are enabled
Summary:
Looks like wedge100's sensor.c is the same as the common code
except that it supports 2 out of 5 of the supported discrete sensors.

This change adds compile time feature selection for these in order
to remove the duplicated code in wedge100.

Test Plan:
Build wedge100.
Used temporary changes using #error to test if the required
feature flags are enabled with others disabled.

Reviewed By: mikechoifb

Differential Revision: D13664684

fbshipit-source-id: 295c37e53