delta psu upgrade script improvements
Summary:
# Always start with a bootloader reset command:

Without this orchestration has to notice PSUs stuck in bootloader when
retrying updates and do a different sequence -- this makes the
orchestration more complex than needed, where doing this here is much
simpler and accomplishes the same thing.

Additionally make sure to *print* exceptions when we fail, not just save
them to the status file.

# Support for saving a transcript of modbus commands that were run

for debugging

# Continue update in the face of CRC check errors reading acknowledgement of blocks written.

See the big comment I added -- long story short this is probably a
workaround for a h/w or firmware issue, haven't fully figured it out.

Will roll this into the backport pkg once it lands in OpenBMC.

Test Plan:
Run upgrades, retry them when they fail, the retries should not fail the
handshake but should at least always get to flashing (in my testing they
do)

Confirmed that verify and reset works and we succesfully get the new FW
rev booted on upgrades that complete when allowing corrupted acks.

Reviewed By: covracer

Differential Revision: D13313928

fbshipit-source-id: 197512b2a