cmock - A poor man's C mocking library for unit-tests
Summary:
Add a simple C-mocking library.
The test code should make sure not to link in the dependent libraries.
There is currently no way to mock out already linked-in functions.
Thus in effect this might be more useful in implementing high-level
unit-tests and functional tests for libraries.

Test Plan: Change aggregate-sensor to use this instead of implementing its own.

Reviewed By: sammyblueeyes

Differential Revision: D14092632

fbshipit-source-id: 558e2eb04e