wedge_power.sh reset -s implementation
Summary:
Power cycling the whole system sometimes failed from CPLD, another method is to do the power cycle by resetting the power sequencer so that the
	 whole board could power cycle. I tested the galaxy and it's already working. I added the change so that the wedge100 can work as well.

Test Plan: Testing is done by commenting resetting through the CPLD. P61031009 is for the galaxy where it is already working. P61030820 is the testing logs for the wedge100-v26.

Reviewed By: vineelasmile

Differential Revision: D14311840

fbshipit-source-id: 1405b09325