Yamp seutil rest api implementation endpoint.
Summary: This is the seutil rest api endpoint implementation for yamp. This is needed by dublin team (Gerardo team)

Test Plan: P61573969 is the paste for the test plan that shows the rest api call working. Some fields are empty, but I have gotten confirmation from both Mike and Gerardo that this is acceptable and meet their needs.

Reviewed By: tomrepo

Differential Revision: D14525743

fbshipit-source-id: a50e15e4e6