Pimserial rest API cache clean up
Summary: When we run pimserial endpoint in rest api, it cached a file for subsequent run where it reads data from it. This file stayed there even when we insert a new pimserial. Therefore, inserting a new pimserial file would read the wrong data because that file was being cached, This task solve that issue by detecting when a pimserial is removed, and if the file exist, it will remove the file so that inserting new pim don't end up reading wrong data.

Test Plan:
Since the clean up will only get trigger when a pim has been
unplug, we have to make sure that we only write to /tmp/ when a pim is
inserted. Therefore, some modification was made in dump_pim_serials and
tested here P64297957. This other paste test and make sure this
functionality is working properly P64286981

Reviewed By: tomrepo

Differential Revision: D15307789

fbshipit-source-id: fc9372c805