openbmc_gpio: Fix potential glitch in gpio line during BMC boot
Summary:
Lightning observed an issue at BMC boot where the GPIO lines toggled
from high to low and then back to high when we were setting the value
(with change_direction=True) to High at boot-up (GPIO initialization).
This temporary trigger to low unfortunately can reset the SSDs.

This change make use of the set-init-value command of gpiocli
to set the value while changing the direction to high thus fixing
the glitch.

Test Plan: Fix is being tested on Lightning.

Reviewed By: tomrepo

Differential Revision: D15339170

fbshipit-source-id: 90eead99d2