openrack: yield in rackmond monitoring loop
Summary:
This prevents the command thread from being starved from being able to
take the lock on the rs485 device, so that external processes aren't
prevented from running commands in a timely manner by the monitoring
loop. That is, we need to give other threads a chance to try for the
lock instead of immediately trying to take it again after releasing it.

Test Plan:
Ran `for i in $(seq 100); <basic modbus command>` on the BMC and timed
it, after this change we can consistently run modbus commands without
seeing long (>10s!) pauses.

{P65388931}

Differential Revision: D15637635

fbshipit-source-id: 81233814c5