CPU consumption test
Summary:
This test case is for the cpu consumption. If it goes above the
expected value, the test will fail. Since cpu consumption for some
needed processes can be quite unpredictable, I have excluded the needed
processes that tends to vary a lot based on cpu consumption.

Test Plan:
Due to the unpredicted nature of the cpu consumption of some
proccesses. I have dumbed lots of log in a user friendly readable way if
the test fail. I have forced it to fail by changing the expected cpu
consumption and (used print to test in the bmc) so that I can see all
the logs. Here is a paste P66494961 of how it will look when it fails.

Here is a paste of that show that it works (I put the expected cpu back
to 60 which is what I expected for wedge40) P66495741

Reviewed By: vineelasmile

Differential Revision: D15829841

fbshipit-source-id: aa45684b97