Adding new logic to showtech for cpu_control and shutdown
Summary:
We had issues of high temperature causing userver to shutdown,
but through the debugging and log collecting process, showtech is a key
utility that we used to collect logs. As of right now, showtech doesn't
provide any information as related to cpu shutdown. This enhancement is
needed to provides us more debugging data when running showtech so that
we can debug quicker.

Test Plan:
Please see this paste for output of detailed test cases
P70317427

Reviewed By: vineelasmile

Differential Revision: D16153672

fbshipit-source-id: d9328e5c81