fby2: only setup fan config on BMC startup (#887)
Summary:
As title
Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/887

Test Plan:
Since it needs few seconds to set pwm from 100% to 70% for detecting whether fans are 10k or 15k fans or not. (If pwm are 70% and fan speeds are stable, rpm >8000 are 15k fans , rpm <8000 are 15k fans)
No need to delay few seconds to set up fan config after sled in.

Reviewed By: benwei13

Differential Revision: D16121547

fbshipit-source-id: af28ece4f0