send vendor information in DHCP requests (whitelisted for wedge only)
Summary:
This introduces the possibility of sending device information in DHCP requests.

## What will we be sending?

A string, with the following proposed format:

```
OpenBMC:key1=value1:key2=value2:[...:keyN=valueN]
```
This is similar to what, for example, Arista uses:
```
Arista;MODEL;VERSION;SERIAL
```
But more flexible, since there are no predefined fields.

For now, for FB platforms, the fields we plan to use are:
`model`: Hardware model
`serial`: Serial Number
`side`: (Only for Galaxy / Backpack) For Line Cards where multiple BMCs share the same serial number, the side (`LEFT` or `RIGHT`)

## How will this be included in the DHCP packets?

We will include this the same way multiple network vendors (Arista, ZPE, Juniper) do, for ZTP.

### DHCPv4

This will be included in the "Vendor Class Identifier" option (60).

### DHCPv6

This will be included as a sub-option in the "Vendor-specific Information Option" (17). The sub-option number will be 1.
See https://tools.ietf.org/html/rfc3315#page-86.
For this option, an Enterprise Number (https://www.iana.org/assignments/enterprise-numbers/enterprise-numbers) needs to be included to be compliant.

## How will this be configured in the BMC?

This is set via a script in `if-pre-up.d`, added to all platforms, that calls `/usr/local/bin/dhcp-id`, which should contain the platform-specific code to fetch the information from the hardware and produce the string. It will also call `/usr/local/bin/enterprise-num` to get the IANA Enterprise Number.

This commit also adds a version of `/usr/local/bin/dhcp-id` that is compatible with all FB Network hardware (along with a script that outputs FB's Enterprise Number), and adds it to Wedge only, to start testing the feature. Rolling it out to other platforms is just a matter of adding the module.

Test Plan:
Flashed image into a wedge, this ran successfully on boot:
```
Configuring network interfaces... Setting vendor information in /etc/dhcp/dhclient.conf
```

- DHCPv4 packets contain option 60 set, to `OpenBMC:model=Wedge-DC-F:serial=<serial>`
- DHCPv6 packets contain option 17 set, with FB's enterprise number (40981), with a single sub-option (code 1), containing the string above.

Reviewed By: sammyblueeyes

Differential Revision: D16259932

fbshipit-source-id: 7a0506946a