Add filter for inlet_temp
Summary:
We have one instance where the inlet_temp of the yamp goes to
86 deg C. Even though, we have applied a patch to the kernel driver of
the chip, we are still seeing such temperature reading which is very
high above the threshold. After some talk with our thermal engineer, it
was concluded that it is safe to add such a filter.

Test Plan:
To test this, I temporarily change the limit value in the
json file to 20 to force this to end up reading a value higher than the
expected temperature. Then I also comment out the shutdown function call
because I didn't want it to shutdown the userver. The following paste
P75610814 shows where it keep printing all those inlet_temp/counter value,
the counter gets reset to 0 if you read a value that isn't above the threshold.

Reviewed By: vineelasmile

Differential Revision: D16557093

fbshipit-source-id: bd242338eb