Extend Rest-piminfo to match Minipack
Summary:
FBOSS PE requested from us to enhance rest-piminfo so that it
can match the one from minipack. Since they are using the same tool,
it's important for the endpoint to be consistent.

Test Plan:
By looking at the original version, I could see that the
value part of the dictionary is a string. Out of curiosity, I still copy
and tested it. Please see this P79390581.

Then, I tested the endpoint using the curl command to make sure that it
is working as expected. Please see this P79394103.

There is no yamp rest-api CIT, so I didn't test those.

Reviewed By: mikechoifb

Differential Revision: D16778290

fbshipit-source-id: 1fc078a148