Get piminfo version to be consistent with yamp
Summary:
We lately received request from FBOSS PE to make some changes
to yamp rest_piminfo. We also have to do the same thing to Minipack so
that we can be consistent accross all of our platform. The version is
now in the following format: <Product Production State>.<Product
Version>.<Product sub_version>

Test Plan:
For our testcases, we ran peutil to take a look at the data,
and the we use the curl command to call the piminfo rest endpoint and
verified that the version is as expected. Then we also ran all the cit
test cases for minipack to make sure that everything is working as
expected (all tests succeeded). Please this paste for more details
P79421224

Reviewed By: mikechoifb

Differential Revision: D16779883

fbshipit-source-id: 58c313b025