i2c: aspeed: add slave inactive timeout support
Summary:
In case of multi-master environment, if a peer master incorrectly handles
a bus in the middle of a transaction, I2C hardware hangs in slave state
and it can't escape from the slave state, so this commit adds slave
inactive timeout support to recover the bus in the case.

Signed-off-by: Jae Hyun Yoo <jae.hyun.yoo@linux.intel.com>

Test Plan: Booted up minipack-5.0 and verified i2c-0 stuck issue is fixed.

Reviewed By: mikechoifb

Differential Revision: D17274179

fbshipit-source-id: 4a79aa5de4