BIC: Add common library and utility
Summary:
This just copies the source from meta-facebook/meta-fby2/recipes-fby2/plat-*/files/{bic|bic-util}
into common and has no functional changes. The only changes is the rename of the recipe file to
mark this as v0.2.

It is not expected to compile or be functional, but a common framework to start
development of a common BIC library.

All platforms already prefer 0.1 version of libbic,bic-util.
So, this should not affect any platform in a negative way.

Test Plan: All CIT should pass build.

Reviewed By: tomrepo

Differential Revision: D17141685

fbshipit-source-id: 161a9f8aa4